A couple of optimizations that seem to pretty dramatically improve performance.  One of the problems I encountered with a fairly complex parser was that it was spending a lot of time formatting error messages in Atom constructors that were never used during successful parsing.  Because the error messages can recursively include string representations of other Atoms, this got expensive fast.  By moving them out into their own methods, we can delay evaluation until they are needed and also memoize the result so we only do it once.