@olbrich 2018-02-07
@kschiess can you take a look at this?