@kschiess 2018-02-07
I can and I will, just on my own time scale. This is open source. 

Just a heads up: Your diff contains lots of changes that are unrelated to what you describe. This makes it a) harder for me to take a quick look and it will b) mess up our git history. I like focused PRs.