@olbrich 2018-02-07
@kschiess Much appreciated.  Just wanted to make sure you were aware of the PR. Looks like I got a bit carried away with rubocop / trailing whitespace autocorrection. I'll update the PR with a tighter change set shortly.