@olbrich 2018-02-10
@kschiess I've cleaned up this PR to keep it to just the essential changes.

For reference:
on master... 
```
time ruby example/optimized_erb.rb
...
real	0m0.241s
user	0m0.101s
sys	0m0.054s
```

on this branch...
```
time ruby example/optimized_erb.rb
...
real	0m0.117s
user	0m0.081s
sys	0m0.032s
```