@kschiess 2018-02-12
Hei @olbrich, 

Thanks for taking the time to clean this up. I think I understand now. 

I've measured the difference between your performance tweaks and the master of parslet as per today using my own [set of benchmarks](https://github.com/kschiess/parslet-benchmarks). Here are the results: 

@master

```
Using parslet at /Users/kschiess/git/own/parslet/lib@7ac42d9bbd92345e3decf891256097debe51720b
benchmarks/001-treetop.rb                 0.152259   0.015071   0.167330 (  0.170218)
benchmarks/002-http_parser.rb             0.692536   0.045098   0.737634 (  0.739403)
benchmarks/003-smalltalk.rb               2.525067   0.185219   2.710286 (  2.710744)
benchmarks/004-csv.rb                     3.396876   2.547524   5.944400 (  5.945829)
benchmarks/005-minip.rb                   0.936701   0.004827   0.941528 (  0.941517)
```

@PR #185

```
Using parslet at /Users/kschiess/git/own/parslet/lib@c0d3d8717c3a49bdae9b83b8239c88f60e843ea6
benchmarks/001-treetop.rb                 0.128832   0.015148   0.143980 (  0.145210)
benchmarks/002-http_parser.rb             0.723474   0.048377   0.771851 (  0.775632)
benchmarks/003-smalltalk.rb               2.593355   0.172649   2.766004 (  2.766053)
benchmarks/004-csv.rb                     3.528174   2.572153   6.100327 (  6.100973)
benchmarks/005-minip.rb                   1.008138   0.005431   1.013569 (  1.013840)
```

All of this using Ruby 2.5.0 on a fairly recent Mac Pro. 

The 'optimized_erb' sample yields these: 

```
master: 		0.080000   0.000000   0.080000 (  0.076035)
your branch:  	0.070000   0.000000   0.070000 (  0.074347)
```

All in all I don't get a clear impression from this. My performance timers are unprecise, certainly, but the benchmark suite does many repetitions to get bigger numbers. 

What do you think?