@olbrich 2018-02-12
@kschiess Thanks for taking a look.  I think the particular cases that will benefit the most from these improvements are ones where you can get deeply nested parsers.    

This all started when I began to rewrite # olbrich/ruby-units to use parslet to parse unit expressions.  On my machine (using ruby 2.5.0) the ruby-units refactor-parser branch takes ~15.8 seconds to run through a test suite of 963 examples.  With this PR, the same tests take about ~3 seconds.  My guess is that there is something that I'm doing in my parser that isn't covered by your benchmarks.

In any event, the parser I'm trying to optimize is [here..](https://github.com/olbrich/ruby-units/blob/refactor-parser/lib/ruby_units/parsers/standard.rb) if you want to take a look.  It's fairly complex, and perhaps there are some things I can do to make it inherently more efficient.  Feedback on that is welcome.