@kschiess 2018-02-13
Hi again, 

I've taken the liberty of adding your parser to the benchmarks. This is what I get: 

Current master: 

```
Using parslet at /Users/kschiess/git/own/parslet/lib@7ac42d9bbd92345e3decf891256097debe51720b
benchmarks/001-treetop.rb                 0.160000   0.020000   0.180000 (  0.176386)
benchmarks/002-http_parser.rb             0.730000   0.050000   0.780000 (  0.781952)
benchmarks/003-smalltalk.rb               2.650000   0.170000   2.820000 (  2.816006)
benchmarks/004-csv.rb                     3.540000   2.050000   5.590000 (  5.587017)
benchmarks/005-minip.rb                   1.010000   0.000000   1.010000 (  1.018649)
benchmarks/006-ruby_units.rb              4.930000   0.430000   5.360000 (  5.364590)
```

Your PR: 

```
Using parslet at /Users/kschiess/git/own/parslet/lib@c0d3d8717c3a49bdae9b83b8239c88f60e843ea6
benchmarks/001-treetop.rb                 0.130000   0.010000   0.140000 (  0.139433)
benchmarks/002-http_parser.rb             0.700000   0.040000   0.740000 (  0.755331)
benchmarks/003-smalltalk.rb               2.660000   0.160000   2.820000 (  2.824107)
benchmarks/004-csv.rb                     3.480000   2.080000   5.560000 (  5.565576)
benchmarks/005-minip.rb                   1.030000   0.000000   1.030000 (  1.040488)
benchmarks/006-ruby_units.rb              4.730000   0.460000   5.190000 (  5.184958)
```



This is more for completeness sake; if the changes are important to you, I'll merge them, since they don't have a negative effect. Thank you for this contribution. I enjoyed this. 

kaspar