@kschiess 2018-02-14
I've released this as '1.8.2'. Enjoy!